+++ astropy/cosmology/io/__init__.py
@@ -6,4 +6,4 @@ Read/Write/Interchange methods for `astropy.cosmology`. **NOT public API**.
 """
 
 # Import the interchange to register them into the io registry.
-from astropy.cosmology._io import mapping  # noqa: F403
+from . import mapping  # noqa: F403
