+++ astropy/cosmology/io/mapping.py
@@ -10,16 +10,16 @@ from astropy.io import registry as io_registry
 from astropy.table import QTable
 
 
-def from_mapping(map, *, move_to_meta=False, **kwargs):
+def from_mapping(map, *, move_to_meta=False, cosmology=None):
     """Load `~astropy.cosmology.Cosmology` from mapping object.
 
     Parameters
     ----------
     map : mapping
-        Must have field "cosmology" which can be either the string name of the
-        cosmology class (e.g. "FlatLambdaCDM") or the class itself.
-        The remaining fields are arguments into the class --
-        like "name" or "meta".
+        Arguments into the class -- like "name" or "meta".
+        If 'cosmology' is None, must have field "cosmology" which can be either
+        the string name of the cosmology class (e.g. "FlatLambdaCDM") or the
+        class itself.
 
     move_to_meta : bool (optional, keyword-only)
         Whether to move keyword arguments that are not in the Cosmology class'
@@ -29,9 +29,11 @@ def from_mapping(map, *, move_to_meta=False, **kwargs):
         preferring specified metadata in the case of a merge conflict
         (e.g. for ``Cosmology(meta={'key':10}, key=42)``, the ``Cosmology.meta``
         will be ``{'key': 10}``).
-    **kwargs
-        Not used.
-        If 'format' is a kwarg, it must be 'mapping'.
+
+    cosmology : str, `~astropy.cosmology.Cosmology` class, or None (optional, keyword-only)
+        The cosmology class (or string name thereof) to use when constructing
+        the cosmology instance. The class also provides default parameter values,
+        filling in any non-mandatory arguments missing in 'map'.
 
     Returns
     -------
@@ -44,7 +46,8 @@ def from_mapping(map, *, move_to_meta=False, **kwargs):
     :meth:`~astropy.cosmology.Cosmology.to_format`.
 
         >>> from astropy.cosmology import Cosmology, Planck18
-        >>> cm = Planck18.to_format('mapping'); cm
+        >>> cm = Planck18.to_format('mapping')
+        >>> cm
         {'cosmology': <class 'astropy.cosmology.core.FlatLambdaCDM'>,
          'name': 'Planck18', 'H0': <Quantity 67.66 km / (Mpc s)>, 'Om0': 0.30966,
          'Tcmb0': 2.7255, 'Neff': 3.046, 'm_nu': <Quantity [0. , 0. , 0.06] eV>,
@@ -57,17 +60,27 @@ def from_mapping(map, *, move_to_meta=False, **kwargs):
         >>> cosmo
         FlatLambdaCDM(name="Planck18", H0=67.7 km / (Mpc s), Om0=0.31,
                       Tcmb0=2.725 K, Neff=3.05, m_nu=[0. 0. 0.06] eV, Ob0=0.049)
+
+    Specific cosmology classes can be used to parse the data. The class'
+    default parameter values are used to fill in any information missing in the
+    data.
+
+        >>> from astropy.cosmology import FlatLambdaCDM
+        >>> del cm["Tcmb0"]  # show FlatLambdaCDM provides default
+        >>> FlatLambdaCDM.from_format(cm)
+        FlatLambdaCDM(name="Planck18", H0=67.7 km / (Mpc s), Om0=0.31,
+                      Tcmb0=0 K, Neff=3.05, m_nu=None, Ob0=0.049)
+
     """
     params = copy.deepcopy(map)  # so can pop
 
     # get cosmology
     # 1st from 'kwargs'. Allows for override of the cosmology, if on file.
     # 2nd from params. This MUST have the cosmology if 'kwargs' did not.
-    if "cosmology" in kwargs:
-        cosmology = kwargs.pop("cosmology")
-        params.pop("cosmology", None)  # pop, but don't use
-    else:
+    if cosmology is None:
         cosmology = params.pop("cosmology")
+    else:
+        params.pop("cosmology", None)  # pop, but don't use
     # if string, parse to class
     if isinstance(cosmology, str):
         cosmology = _COSMOLOGY_CLASSES[cosmology]
@@ -87,30 +100,29 @@ def from_mapping(map, *, move_to_meta=False, **kwargs):
     elif move_to_meta:  # prefers current meta, which was explicitly set
         meta = ba.arguments["meta"] or {}  # (None -> dict)
         ba.arguments["meta"] = {**params, **meta}
-    elif bool(params):
+    elif params:
         raise TypeError(f"there are unused parameters {params}.")
     # else: pass  # no kwargs, no move-to-meta, and all the params are used
 
     return cosmology(*ba.args, **ba.kwargs)
 
 
-def to_mapping(cosmology, *args, **kwargs):
+def to_mapping(cosmology, *args):
     """Return the cosmology class, inputs, and metadata as a dict.
 
     Parameters
     ----------
     cosmology : `~astropy.cosmology.Cosmology` subclass instance
+    *args
+        Not used. Needed for compatibility with
+        `~astropy.io.registry.UnifiedReadWriteMethod`
 
     Returns
     -------
     dict
         with key-values for the cosmology parameters and also:
-
         - 'cosmology' : the class
         - 'meta' : the contents of the cosmology's metadata attribute
-    *args, **kwargs
-        Not used. Needed for compatibility with
-        `~astropy.io.registry.UnifiedReadWriteMethod`
 
     Examples
     --------
